wip: LedgerState next
