+++ cnd/src/swap_protocols/han/ethereum/events.rs
@@ -164,110 +164,3 @@ async fn htlc_refunded<A: Asset>(
 
     Ok(Refunded { transaction })
 }
-
-mod erc20 {
-    use super::*;
-    use crate::ethereum::U256;
-    use asset::ethereum::FromWei;
-
-    #[async_trait::async_trait]
-    impl HtlcFunded<Ethereum, asset::Erc20> for Cache<Web3Connector> {
-        async fn htlc_funded(
-            &self,
-            htlc_params: HtlcParams<Ethereum, asset::Erc20>,
-            htlc_deployment: &Deployed<Ethereum>,
-            start_of_swap: NaiveDateTime,
-        ) -> anyhow::Result<Funded<Ethereum, asset::Erc20>> {
-            let connector = self.clone();
-            let events = Some(vec![Event {
-                address: Some(htlc_params.asset.token_contract),
-                data: None,
-                topics: vec![
-                    Some(Topic(*super::TRANSFER_LOG_MSG)),
-                    None,
-                    Some(Topic(htlc_deployment.location.into())),
-                ],
-            }]);
-            let TransactionAndReceipt {
-                transaction,
-                receipt,
-            } = matching_transaction(
-                connector,
-                TransactionPattern {
-                    from_address: None,
-                    to_address: None,
-                    is_contract_creation: None,
-                    transaction_data: None,
-                    transaction_data_length: None,
-                    events,
-                },
-                start_of_swap,
-            )
-            .await
-            .context("failed to find transaction to fund htlc")?;
-            let log = receipt
-                .logs
-                .into_iter()
-                .find(|log| log.topics.contains(&*super::TRANSFER_LOG_MSG))
-                .expect("Fund transaction receipt must contain transfer events");
-
-            let quantity =
-                asset::Erc20Quantity::from_wei(U256::from_big_endian(log.data.0.as_ref()));
-            let asset = asset::Erc20::new(log.address, quantity);
-
-            Ok(Funded { transaction, asset })
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl HtlcDeployed<Ethereum, asset::Erc20> for Cache<Web3Connector> {
-        async fn htlc_deployed(
-            &self,
-            htlc_params: HtlcParams<Ethereum, asset::Erc20>,
-            start_of_swap: NaiveDateTime,
-        ) -> anyhow::Result<Deployed<Ethereum>> {
-            let connector = self.clone();
-            let pattern = TransactionPattern {
-                from_address: None,
-                to_address: None,
-                is_contract_creation: Some(true),
-                transaction_data: Some(htlc_params.bytecode()),
-                transaction_data_length: None,
-                events: None,
-            };
-            let TransactionAndReceipt { transaction, .. } =
-                matching_transaction(connector, pattern, start_of_swap)
-                    .await
-                    .context("failed to find transaction to deploy htlc")?;
-
-            Ok(Deployed {
-                location: calculate_contract_address_from_deployment_transaction(&transaction),
-                transaction,
-            })
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl HtlcRedeemed<Ethereum, asset::Erc20> for Cache<Web3Connector> {
-        async fn htlc_redeemed(
-            &self,
-            htlc_params: HtlcParams<Ethereum, asset::Erc20>,
-            htlc_deployment: &Deployed<Ethereum>,
-            start_of_swap: NaiveDateTime,
-        ) -> anyhow::Result<Redeemed<Ethereum>> {
-            htlc_redeemed(self.clone(), htlc_params, htlc_deployment, start_of_swap).await
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl HtlcRefunded<Ethereum, asset::Erc20> for Cache<Web3Connector> {
-        async fn htlc_refunded(
-            &self,
-            htlc_params: HtlcParams<Ethereum, asset::Erc20>,
-            htlc_deployment: &Deployed<Ethereum>,
-            start_of_swap: NaiveDateTime,
-        ) -> anyhow::Result<Refunded<Ethereum>> {
-            htlc_refunded(self.clone(), htlc_params, htlc_deployment, start_of_swap).await
-        }
-    }
-}
