+++ cnd/src/swap_protocols/han/mod.rs
@@ -1,0 +1,2 @@
+pub mod events;
+pub use events::{Funded, Redeemed, Refunded, WatchFunded, WatchRedeemed, WatchRefunded};
