+++ cnd/src/swap_protocols/herc20/mod.rs
@@ -1,0 +1,4 @@
+pub mod events;
+pub use events::{
+    Deployed, Funded, Redeemed, Refunded, WatchDeployed, WatchFunded, WatchRedeemed, WatchRefunded,
+};
