+++ cnd/src/swap_protocols/mod.rs
@@ -1,5 +1,7 @@
 pub mod actions;
 mod facade;
+pub mod han;
+pub mod herc20;
 pub mod ledger;
 pub mod rfc003;
 mod swap_id;
